aramex shipment tracking support feature update
Aramex shipping API integration

Shipping API integration

Shipping API integration + rating api

Shipping and rating api integration aramex

Shipping and rating api integration aramex

Shipping api changes

Shipping api changes

Shipping api changes

Shipping api integration

Shipping api integration

Aramex api integration - Integration with new credentials given by Aramex support

Resolved errors

Aramex WSDL changes

Resolved conflicts

Resolved package type error

Resolved package type error

Aramex shipping api integration

Aramex shipping api integration

Shipping api integration

Modified source code as per reviewers comments

Aramex api changes

Shipping and rating API changes Aramex

Shipping and rating API changes Aramex

Shipping and rating API changes Aramex

Aramex shipping APi changes

Aramex shipping APi changes

Aramex shipping APi changes

Aramex shipping api - Made changes as per the reviewers comments

Added OND to the product type

Tests for aramex shipping api

Aramex shipping API enhancement

Aramex shipping API enhancement and Tests

Changes as per reviewers comments

Made changes as per the reviewers comments - 2 left

Made changes as per the reviewers comments - 2 left

Revised code for remaining 2 review comments

Revised code for remaining 2 review comments - round 2

Revised code for remaining 2 review comments - round 3

Revised code for remaining 2 review comments - round 4

Shipment item data restructure

Shipment item data restructure 1

Modified shimpmet request processing section as per the new WSDL

Modified shimpmet request processing section as per the new WSDL

Shipment item structure change

Convert report type to byte stream and byte stream to pdf and written tests for checking declarations

Whitlist conuntries

Indentation

Indentation and altered service list

Added product type

Loaded whitelist through setting

Loaded whitelist through setting - added comments

Load whitelist country

Load whitelist country

dopy of carrier list

copy of carrier list

Removed unneccessary import

+++ postal/postal.py
@@ -25,7 +25,7 @@ class Postal:
         """
         # Give carriers a back reference to the main postal object.
         configuration_dict['postal'] = self
-
+        self.carrier_country = configuration_dict['carrier_country']
         self.carriers = {carrier.name: carrier
                          for carrier in configuration_dict['enabled_carriers']}
         carrier_configs = configuration_dict['carrier_inits']
@@ -34,6 +34,7 @@ class Postal:
                 self.carriers[name] = carrier(
                     postal_configuration=configuration_dict,
                     **carrier_configs[name])
+                self.carrier_list = self.carriers
             except:
                 print 'Error while constructing carrier ' + str(name)
                 print 'with these args: ' + str(carrier_configs[name])
@@ -65,13 +66,20 @@ class Postal:
                 raise NotSupportedError('The dimensions of package #%s are '
                                         'invalid.' % i)
 
-        thread_pool = ThreadPool(processes=len(self.carriers))
+        # Check country white list
+        for carrier in self.carrier_list.values():
+            served = get_served_country(carrier.name, request.destination.country.alpha2, self.carrier_country)
+            if not served:
+                del self.carrier_list[carrier.name]
 
+        thread_pool = ThreadPool(processes=len(self.carrier_list))
         result = dict(thread_pool.map(
-            _task, [(carrier, request) for carrier in self.carriers.values()]))
-        thread_pool.terminate()
-        thread_pool.join()
-        return result
+            _task, [(carrier, request) for carrier in self.carrier_list.values()]))
+        self.carrier_list = self.carriers
+        if result:
+            thread_pool.terminate()
+            thread_pool.join()
+            return result
 
     def get_all_services(self):
         """
@@ -118,6 +126,7 @@ class Postal:
 
 
 def _task(arg_list):
+
     carrier, request = arg_list
     data_dict = {'services': None, 'error': None}
     try:
@@ -128,3 +137,9 @@ def _task(arg_list):
         data_dict['error'] = err
 
     return carrier, data_dict
+
+
+def get_served_country(carrier, dest_country, carrier_country):
+    if carrier in carrier_country and not dest_country in carrier_country[carrier]:
+        return False
+    return True
